+++ file-search/contents/ui/components/NativeContextMenuSeparator.qml
@@ -1,0 +1,5 @@
+import org.kde.plasma.extras as PlasmaExtras
+
+PlasmaExtras.MenuItem {
+    separator: true
+}